Fix signatures

+++ includes/utils.h
@@ -1,9 +1,9 @@
 #include "quickjs.h"
 
-static int JSToInt32(JSContext *ctx, JSValueConst val);
+static int32_t JSToInt32(JSContext *ctx, JSValueConst val);
 
-static int JSToInt64(JSContext *ctx, JSValueConst val);
+static int64_t JSToInt64(JSContext *ctx, JSValueConst val);
 
-static int JSToInt64Ext(JSContext *ctx, JSValueConst val);
+static int64_t JSToInt64Ext(JSContext *ctx, JSValueConst val);
 
-static int JSToFloat64(JSContext *ctx, JSValueConst val);
+static double JSToFloat64(JSContext *ctx, JSValueConst val);
